requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.7.2
pandas==2.2.2
python-dotenv==1.0.1
pydantic==2.8.2
sqlalchemy==2.0.35
pymysql==1.1.1
psycopg2-binary==2.9.10
pyodbc==5.1.0
scikit-learn==1.3.2
networkx==3.2.1
python-multipart==0.0.6